readonly file 처리방법

+++ Nado Game/pyqt5/사진관리/pictures_control.py
@@ -1,5 +1,6 @@
 import sys
 import os
+import stat
 import pathlib
 import pandas as pd
 from PyQt5.QtCore import Qt, pyqtSlot, QObject, pyqtSignal
@@ -250,9 +251,15 @@ class ElWindow(QMainWindow, form_class):
         # 사진찍은 날짜 가져오기
         t_time = self.takePictureTime(path, f)
         filename = os.path.join(path, f)
+        T = os.stat(filename)
+        print('T.st_ctime, T.st_mtime, T.st_atime', T.st_ctime, T.st_mtime, T.st_atime)
+        [c_time, m_time, a_time] = [T.st_ctime, T.st_mtime, T.st_atime]
+        print(c_time, m_time, a_time)
         c_time = os.path.getctime(filename)
         m_time = os.path.getmtime(filename)
         a_time = os.path.getatime(filename)
+        
+        print(c_time, m_time, a_time)
         min_time = min(t_time, c_time, m_time, a_time)
         dt = datetime.fromtimestamp(min_time)
         y = dt.strftime("%Y"+l[0])
@@ -307,7 +314,7 @@ class ElWindow(QMainWindow, form_class):
                 C_files += 1
                 shutil.copy2(f, t) # 파일 복사 (파일 개정 시간 등 포함하여 복사를 위해 copy2 사용)pass
                 self.disp_C_files(C_files)
-                print(C_files)
+                #print(C_files)
                 self.listWidget_2.addItem(str(folder[0]) +' ' + str(t) +' ' + folder[4])
                 CFile.append([folder[0], t, folder[4]])
         self.removeTempFile()
@@ -341,7 +348,12 @@ class ElWindow(QMainWindow, form_class):
             t_file = pathlib.Path(t, folder[4]) # target file 경로 및 이름 
             t.mkdir(parents=True, exist_ok=True) # 파일 경로에 있는 모든 폴더를 생성함. 있으면 놔둠
             if os.path.isfile(pathlib.Path(t_file)):
-                os.remove(f)
+                if os.stat(f).st_mode == 33060: # 33060 readonly, 33206 writable
+                    try:
+                        os.chmod(f, stat.S_IWRITE)
+                        os.remove(f)
+                    except:
+                        os.remove(f)
                 R_files += 1
                 self.lineEdit_9.setText(str(R_files))
                 self.listWidget_4.addItem(str(folder[0]) + '/' + folder[4])
